refactor: move helpers to helper namespace
