+++ src/ReturnTypes/EloquentBuilderExtension.php
@@ -8,9 +8,9 @@ use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Str;
 use NunoMaduro\Larastan\Concerns;
-use NunoMaduro\Larastan\Methods\BuilderHelper;
+use NunoMaduro\Larastan\Helpers\BuilderHelper;
 use NunoMaduro\Larastan\Methods\ModelForwardsCallsExtension;
-use NunoMaduro\Larastan\Methods\ModelTypeHelper;
+use NunoMaduro\Larastan\Helpers\ModelTypeHelper;
 use PhpParser\Node\Expr\MethodCall;
 use PHPStan\Analyser\Scope;
 use PHPStan\Reflection\BrokerAwareExtension;
