+++ src/ReturnTypes/ModelExtension.php
@@ -9,8 +9,8 @@ use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use NunoMaduro\Larastan\Concerns;
-use NunoMaduro\Larastan\Methods\BuilderHelper;
-use NunoMaduro\Larastan\Methods\ModelTypeHelper;
+use NunoMaduro\Larastan\Helpers\BuilderHelper;
+use NunoMaduro\Larastan\Helpers\ModelTypeHelper;
 use NunoMaduro\Larastan\Methods\Pipes\Mixins;
 use PhpParser\Node\Expr\StaticCall;
 use PHPStan\Analyser\Scope;
